SintraPrime_Mythic_Speech/requirements.txt
requests
python-dotenv
flask
slack_sdk
